Make sure VRCX stuff only happens on Windows
