+++ src/vrcx/unix.rs
@@ -1,0 +1,11 @@
+use super::VrcxStartup;
+
+impl VrcxStartup {
+    pub fn vrcx_installed(&self) -> bool {
+        false
+    }
+
+    pub fn shortcut_exists(&self) -> bool {
+        false
+    }
+}
